backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
msal
openai
azure-ai-formrecognizer
